moved methods to utils
